+++ lib/antwort.rb
@@ -4,6 +4,7 @@ end
 
 require_relative 'antwort/builder'
 require_relative 'antwort/cli'
+require_relative 'antwort/cli/upload'
 require_relative 'antwort/server/assets'
 require_relative 'antwort/server/helpers'
 require_relative 'antwort/server'