+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 * Added `antwort` CLI
   * `$ antwort init MY_PROJECT` initializes the project 
   * `$ antwort new FOO_LETTER` sets up a new email 
+  * `$ antwort upload FOO_LETTER` uploads the assets to Amazon S3
 
 ### Fixed